+++ package.json
@@ -4,12 +4,15 @@
   "description": "JavaScript implementation of the IPLD (InterpPlanetary Linked Data)",
   "main": "src/index.js",
   "scripts": {
-    "test": "./node_modules/.bin/lab tests/*-test.js",
-    "codestyle-fix": "./node_modules/.bin/standard --format",
-    "codestyle": "./node_modules/.bin/standard"
+    "test": "dignified-test",
+    "test:browser": "dignified-test browser",
+    "test:node": "dignified-test browser",
+    "lint": "dignified-lint",
+    "release": "dignified-release"
   },
   "pre-commit": [
-    "codestyle"
+    "lint",
+    "test"
   ],
   "repository": {
     "type": "git",
@@ -28,16 +31,15 @@
   },
   "homepage": "https://github.com/diasdavid/js-ipld",
   "dependencies": {
-    "cbor": "^0.3.14",
+    "cbor": "^1.0.1",
     "cbor-sync": "^1.0.2",
-    "multihashing": "^0.1.3",
+    "multihashing": "^0.2.0",
     "remap-keys": "^1.0.1",
     "underscore": "^1.8.3"
   },
   "devDependencies": {
-    "code": "^1.5.0",
-    "lab": "^5.15.2",
-    "pre-commit": "^1.1.1",
-    "standard": "^5.1.0"
+    "chai": "^3.5.0",
+    "dignified.js": "github:dignifiedquire/dignified.js",
+    "pre-commit": "^1.1.1"
   }
 }
